requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile pyproject.toml -o requirements.txt
alembic==1.17.2
    # via rentory (pyproject.toml)
annotated-doc==0.0.4
    # via fastapi
annotated-types==0.7.0
    # via pydantic
anyio==4.10.0
    # via
    #   httpx
    #   starlette
argon2-cffi==25.1.0
    # via rentory (pyproject.toml)
argon2-cffi-bindings==25.1.0
    # via argon2-cffi
authlib==1.6.5
    # via rentory (pyproject.toml)
certifi==2025.8.3
    # via
    #   httpcore
    #   httpx
cffi==1.17.1
    # via
    #   argon2-cffi-bindings
    #   cryptography
click==8.2.1
    # via uvicorn
cryptography==45.0.6
    # via authlib
dnspython==2.7.0
    # via email-validator
email-validator==2.2.0
    # via pydantic
fastapi==0.121.2
    # via rentory (pyproject.toml)
greenlet==3.2.4
    # via sqlalchemy
h11==0.16.0
    # via
    #   httpcore
    #   uvicorn
httpcore==1.0.9
    # via httpx
httpx==0.28.1
    # via rentory (pyproject.toml)
idna==3.10
    # via
    #   anyio
    #   email-validator
    #   httpx
itsdangerous==2.2.0
    # via rentory (pyproject.toml)
mako==1.3.10
    # via alembic
markupsafe==3.0.3
    # via mako
mysql-connector-python==9.5.0
    # via rentory (pyproject.toml)
phonenumbers==9.0.18
    # via rentory (pyproject.toml)
pillow==12.0.0
    # via rentory (pyproject.toml)
pycparser==2.22
    # via cffi
pydantic==2.12.4
    # via
    #   rentory (pyproject.toml)
    #   fastapi
    #   pydantic-extra-types
    #   pydantic-settings
    #   sqlmodel
pydantic-core==2.41.5
    # via pydantic
pydantic-extra-types==2.10.6
    # via rentory (pyproject.toml)
pydantic-settings==2.12.0
    # via rentory (pyproject.toml)
pyjwt==2.10.1
    # via rentory (pyproject.toml)
python-dotenv==1.1.1
    # via pydantic-settings
python-multipart==0.0.20
    # via rentory (pyproject.toml)
sniffio==1.3.1
    # via anyio
sqlalchemy==2.0.43
    # via
    #   alembic
    #   sqlmodel
sqlmodel==0.0.27
    # via rentory (pyproject.toml)
starlette==0.47.2
    # via fastapi
systemd-python==235
    # via rentory (pyproject.toml)
typing-extensions==4.14.1
    # via
    #   alembic
    #   fastapi
    #   pydantic
    #   pydantic-core
    #   pydantic-extra-types
    #   sqlalchemy
    #   typing-inspection
typing-inspection==0.4.2
    # via
    #   pydantic
    #   pydantic-settings
uvicorn==0.38.0
    # via rentory (pyproject.toml)
